+++ proxy/routehandler/routehandler.go
@@ -1,39 +1,27 @@
 package routehandler
 
 import (
+	"fmt"
 	"net/http"
 
-	"github.com/GoMetric/statsd-http-proxy/proxy/statsdclient"
+	"github.com/InjectiveLabs/statsd-http-proxy/proxy/statsdclient"
 )
 
 // RouteHandler as a collection of route handlers
 type RouteHandler struct {
-	statsdClient statsdclient.StatsdClientInterface
-	metricPrefix string
+	statter statsdclient.Statter
 }
 
 // NewRouteHandler creates collection of route handlers
 func NewRouteHandler(
-	statsdClient statsdclient.StatsdClientInterface,
-	metricPrefix string,
+	statter statsdclient.Statter,
 ) *RouteHandler {
-	// prepare metric prefix
-	if metricPrefix != "" && (metricPrefix)[len(metricPrefix)-1:] != "." {
-		metricPrefix = metricPrefix + "."
-	}
-
 	// build route handler
 	routeHandler := RouteHandler{
-		statsdClient,
-		metricPrefix,
+		statter,
 	}
 
 	return &routeHandler
-}
-
-// GetFullyQualifiedMetricKey return metric key with passed suffix and pre-configured prefix
-func (routeHandler *RouteHandler) getFullyQualifiedMetricKey(metricKeySuffix string) string {
-	return routeHandler.metricPrefix + metricKeySuffix
 }
 
 //HandleMetric reads count, gauge, timing and set metrics from HTTP and sent them to StatsD
@@ -41,20 +29,22 @@ func (routeHandler *RouteHandler) HandleMetric(
 	w http.ResponseWriter,
 	r *http.Request,
 	metricType string,
-	metricKeySuffix string,
+	metricKey string,
 ) {
-	// get fully qualified metric key
-	metricKey := routeHandler.getFullyQualifiedMetricKey(metricKeySuffix)
-
-	// run handler
 	switch metricType {
 	case "count":
 		routeHandler.handleCountRequest(w, r, metricKey)
+	case "incr":
+		routeHandler.handleIncrementRequest(w, r, metricKey)
 	case "gauge":
 		routeHandler.handleGaugeRequest(w, r, metricKey)
 	case "timing":
 		routeHandler.handleTimingRequest(w, r, metricKey)
-	case "set":
-		routeHandler.handleSetRequest(w, r, metricKey)
+	case "uniq":
+		routeHandler.handleUniqueRequest(w, r, metricKey)
 	}
+}
+
+func (routeHandler *RouteHandler) HandleHeartbeatRequest(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprint(w, "OK")
 }