+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0-stable
+  * use `alexcesaro/statsd`
+  * adapt the stats interface
+
 ## 1.0-dev
   * Default HTTP port changed to 8825
   * Configure HTTP timeouts through `--http-timeout-read`, `--http-timeout-write` and `--http-timeout-idle`